Add an abstraction layer for tokens
This patch adds a common interface for different kinds of tokens. The
interface should hopefully be able to support both smart cards and
file-based tokens (PKCS12 and NGE/NGP), but at the moment only PKCS12
tokens are implemented.

I've included a few other changes in this patch as well because it was
easier to make them at the same time. This includes more useful error
messages and a little refactoring of the PKCS12 code.

+++ client/posix.c
@@ -44,7 +44,6 @@
 
 #include "../common/defines.h"
 #include "platform.h"
-#include "misc.h"
 
 void platform_seedRandom() {
     struct timeval tv;
@@ -114,7 +113,7 @@ void platform_closeDir(PlatformDirIter *iter) {
 }
 
 #define NUM_PATHS 2
-void platform_keyDirs(char*** path, int* len) {
+void platform_keyDirs(char*** path, size_t* len) {
     static const char suffix[] = "cbt";
     static const char hidden_suffix[] = ".cbt";
     static char *paths[NUM_PATHS];
